requirements.txt
flask
flasgger
mysql-connector-python
gunicorn
Flask-SQLAlchemy
Flask-Migrate
PyJWT
cryptography
python-dotenv
scikit-learn
pandas
